requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
supabase
cryptography
pydantic
pydantic-settings
python-dotenv
